+++ activerecord/test/cases/locking_test.rb
@@ -10,7 +10,7 @@ require 'models/string_key_object'
 class LockWithoutDefault < ActiveRecord::Base; end
 
 class LockWithCustomColumnWithoutDefault < ActiveRecord::Base
-  set_table_name :lock_without_defaults_cust
+  self.table_name = :lock_without_defaults_cust
   set_locking_column :custom_lock_version
 end